requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
psutil>=5.9.0
requests>=2.31.0
pydantic>=2.0.0
paramiko>=2.8.0
pyyaml>=6.0
